Refactors functions to telegramUtils
